Finished updates based on peer review based on the fixes.
Fix #1 - Code comments are updated.
Fix #2 - Stock Price Impact of COVID-19 method is added.
Fix #3 - We are measuring based on financial quarters. Seasonality is removed.
Fix #4 - Since we only stock prices data we cannot find casulity. Only trends and returns can be observed. The same is kept as a note in proposal.
Fix #5 - Write-up alignment of plot details and then x-axis and y-axis are mentioned later.

Fix Not provided for "Are the weekday returns comparable within a season or a year?"
WeekDays Returns with quarters will checked in box plot or any other plots. Which is already mentioned in earlier write-up.